Refactored API in a separate module
