+++ client-api/src/main/java/io/streamnative/oxia/client/api/AsyncOxiaClient.java
@@ -15,12 +15,14 @@
  */
 package io.streamnative.oxia.client.api;
 
+import io.streamnative.oxia.client.api.exceptions.UnexpectedVersionIdException;
 import java.util.List;
+import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 import lombok.NonNull;
 
-/** Synchronous client for the Oxia service. */
-public interface SyncOxiaClient extends AutoCloseable {
+/** Asynchronous client for the Oxia service. */
+public interface AsyncOxiaClient extends AutoCloseable {
 
     /**
      * Conditionally associates a value with a key if the server's versionId of the record is as
@@ -32,12 +34,13 @@ public interface SyncOxiaClient extends AutoCloseable {
      * @param key The key with which the value should be associated.
      * @param value The value to associate with the key.
      * @param options Set {@link PutOption options} for the put.
-     * @return The result of the put at the specified key.
-     * @throws UnexpectedVersionIdException The versionId at the server did not that match supplied in
+     * @return The result of the put at the specified key. Supplied via a future that returns the
+     *     {@link PutResult}. The future will complete exceptionally with an {@link
+     *     UnexpectedVersionIdException} if the versionId at the server did not that match supplied in
      *     the call.
      */
-    PutResult put(@NonNull String key, byte @NonNull [] value, PutOption... options)
-            throws UnexpectedVersionIdException;
+    @NonNull
+    CompletableFuture<PutResult> put(String key, byte[] value, PutOption... options);
 
     /**
      * Conditionally deletes the record associated with the key if the record exists, and the server's
@@ -47,12 +50,13 @@ public interface SyncOxiaClient extends AutoCloseable {
      *
      * @param key Deletes the record with the specified key.
      * @param options Set {@link DeleteOption options} for the delete.
-     * @return True if the key was actually present on the server, false otherwise.
-     * @throws UnexpectedVersionIdException The versionId at the server did not that match supplied in
-     *     the call.
+     * @return A future that completes when the delete call has returned. The future can return a flag
+     *     that will be true if the key was actually present on the server, false otherwise. The
+     *     future will complete exceptionally with an {@link UnexpectedVersionIdException} the
+     *     versionId at the server did not that match supplied in the call.
      */
-    boolean delete(@NonNull String key, @NonNull DeleteOption... options)
-            throws UnexpectedVersionIdException;
+    @NonNull
+    CompletableFuture<Boolean> delete(String key, DeleteOption... options);
 
     /**
      * Deletes any records with keys within the specified range. For more information on how keys are
@@ -64,8 +68,10 @@ public interface SyncOxiaClient extends AutoCloseable {
      *     the range.
      * @param endKeyExclusive The key that declares the end of the range, and is <b>excluded</b> from
      *     the range.
+     * @return A future that completes when the delete call has returned.
      */
-    void deleteRange(@NonNull String startKeyInclusive, @NonNull String endKeyExclusive);
+    @NonNull
+    CompletableFuture<Void> deleteRange(String startKeyInclusive, String endKeyExclusive);
 
     /**
      * Returns the record associated with the specified key. The returned value includes the value,
@@ -73,8 +79,10 @@ public interface SyncOxiaClient extends AutoCloseable {
      *
      * @param key The key associated with the record to be fetched.
      * @return The value associated with the supplied key, or {@code null} if the key did not exist.
+     *     Supplied via a future returning a {@link GetResult}.
      */
-    GetResult get(@NonNull String key);
+    @NonNull
+    CompletableFuture<GetResult> get(String key);
 
     /**
      * Lists any existing keys within the specified range. For more information on how keys are
@@ -86,11 +94,11 @@ public interface SyncOxiaClient extends AutoCloseable {
      *     the range.
      * @param endKeyExclusive The key that declares the end of the range, and is <b>excluded</b> from
      *     the range.
-     * @return The list of keys that exist within the specified range, or an empty list if there were
-     *     none.
+     * @return The list of keys that exist within the specified range or an empty list if there were
+     *     none. Supplied via a future.
      */
     @NonNull
-    List<String> list(@NonNull String startKeyInclusive, @NonNull String endKeyExclusive);
+    CompletableFuture<List<String>> list(String startKeyInclusive, String endKeyExclusive);
 
     /**
      * Registers a callback to receive Oxia {@link Notification record change notifications}. Multiple
