+++ client-api/src/main/java/io/streamnative/oxia/client/api/SyncOxiaClient.java
@@ -15,13 +15,13 @@
  */
 package io.streamnative.oxia.client.api;
 
+import io.streamnative.oxia.client.api.exceptions.UnexpectedVersionIdException;
 import java.util.List;
-import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 import lombok.NonNull;
 
-/** Asynchronous client for the Oxia service. */
-public interface AsyncOxiaClient extends AutoCloseable {
+/** Synchronous client for the Oxia service. */
+public interface SyncOxiaClient extends AutoCloseable {
 
     /**
      * Conditionally associates a value with a key if the server's versionId of the record is as
@@ -33,13 +33,12 @@ public interface AsyncOxiaClient extends AutoCloseable {
      * @param key The key with which the value should be associated.
      * @param value The value to associate with the key.
      * @param options Set {@link PutOption options} for the put.
-     * @return The result of the put at the specified key. Supplied via a future that returns the
-     *     {@link PutResult}. The future will complete exceptionally with an {@link
-     *     UnexpectedVersionIdException} if the versionId at the server did not that match supplied in
+     * @return The result of the put at the specified key.
+     * @throws UnexpectedVersionIdException The versionId at the server did not that match supplied in
      *     the call.
      */
-    @NonNull
-    CompletableFuture<PutResult> put(String key, byte[] value, PutOption... options);
+    PutResult put(@NonNull String key, byte @NonNull [] value, PutOption... options)
+            throws UnexpectedVersionIdException;
 
     /**
      * Conditionally deletes the record associated with the key if the record exists, and the server's
@@ -49,13 +48,12 @@ public interface AsyncOxiaClient extends AutoCloseable {
      *
      * @param key Deletes the record with the specified key.
      * @param options Set {@link DeleteOption options} for the delete.
-     * @return A future that completes when the delete call has returned. The future can return a flag
-     *     that will be true if the key was actually present on the server, false otherwise. The
-     *     future will complete exceptionally with an {@link UnexpectedVersionIdException} the
-     *     versionId at the server did not that match supplied in the call.
+     * @return True if the key was actually present on the server, false otherwise.
+     * @throws UnexpectedVersionIdException The versionId at the server did not that match supplied in
+     *     the call.
      */
-    @NonNull
-    CompletableFuture<Boolean> delete(String key, DeleteOption... options);
+    boolean delete(@NonNull String key, @NonNull DeleteOption... options)
+            throws UnexpectedVersionIdException;
 
     /**
      * Deletes any records with keys within the specified range. For more information on how keys are
@@ -67,10 +65,8 @@ public interface AsyncOxiaClient extends AutoCloseable {
      *     the range.
      * @param endKeyExclusive The key that declares the end of the range, and is <b>excluded</b> from
      *     the range.
-     * @return A future that completes when the delete call has returned.
      */
-    @NonNull
-    CompletableFuture<Void> deleteRange(String startKeyInclusive, String endKeyExclusive);
+    void deleteRange(@NonNull String startKeyInclusive, @NonNull String endKeyExclusive);
 
     /**
      * Returns the record associated with the specified key. The returned value includes the value,
@@ -78,10 +74,8 @@ public interface AsyncOxiaClient extends AutoCloseable {
      *
      * @param key The key associated with the record to be fetched.
      * @return The value associated with the supplied key, or {@code null} if the key did not exist.
-     *     Supplied via a future returning a {@link GetResult}.
      */
-    @NonNull
-    CompletableFuture<GetResult> get(String key);
+    GetResult get(@NonNull String key);
 
     /**
      * Lists any existing keys within the specified range. For more information on how keys are
@@ -93,11 +87,11 @@ public interface AsyncOxiaClient extends AutoCloseable {
      *     the range.
      * @param endKeyExclusive The key that declares the end of the range, and is <b>excluded</b> from
      *     the range.
-     * @return The list of keys that exist within the specified range or an empty list if there were
-     *     none. Supplied via a future.
+     * @return The list of keys that exist within the specified range, or an empty list if there were
+     *     none.
      */
     @NonNull
-    CompletableFuture<List<String>> list(String startKeyInclusive, String endKeyExclusive);
+    List<String> list(@NonNull String startKeyInclusive, @NonNull String endKeyExclusive);
 
     /**
      * Registers a callback to receive Oxia {@link Notification record change notifications}. Multiple
