+++ client-api/src/main/java/io/streamnative/oxia/client/api/exceptions/OxiaException.java
@@ -13,18 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.streamnative.oxia.client.api;
+package io.streamnative.oxia.client.api.exceptions;
 
-import io.streamnative.oxia.proto.PutResponse;
-import lombok.NonNull;
-
-/**
- * The result of a client get request.
- *
- * @param version Metadata for the record associated with the key specified in the call.
- */
-public record PutResult(@NonNull Version version) {
-    public static @NonNull PutResult fromProto(@NonNull PutResponse response) {
-        return new PutResult(Version.fromProto(response.getVersion()));
+/** A super-class of exceptions describing errors that occurred on an Oxia server. */
+public abstract class OxiaException extends Exception {
+    OxiaException(String message) {
+        super(message);
     }
 }
