+++ client-api/src/main/java/io/streamnative/oxia/client/api/exceptions/SessionDoesNotExistException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.streamnative.oxia.client.api;
+package io.streamnative.oxia.client.api.exceptions;
 
 /** The session does not exist on the server. */
 public class SessionDoesNotExistException extends OxiaException {
