+++ client/src/main/java/io/streamnative/oxia/client/ProtoUtil.java
@@ -15,8 +15,17 @@
  */
 package io.streamnative.oxia.client;
 
+import io.streamnative.oxia.client.api.GetResult;
+import io.streamnative.oxia.client.api.PutResult;
+import io.streamnative.oxia.client.api.Version;
+import io.streamnative.oxia.proto.GetResponse;
+import io.streamnative.oxia.proto.PutResponse;
 import java.nio.ByteBuffer;
+import java.util.Optional;
+import lombok.NonNull;
+import lombok.experimental.UtilityClass;
 
+@UtilityClass
 public class ProtoUtil {
 
     public static int longToUint32(long value) {
@@ -25,5 +34,24 @@ public class ProtoUtil {
 
     public static long uint32ToLong(int unit32AsInt) {
         return ByteBuffer.allocate(8).putInt(0).putInt(unit32AsInt).flip().getLong();
+    }
+
+    public static @NonNull PutResult getPutResultFromProto(@NonNull PutResponse response) {
+        return new PutResult(getVersionFromProto(response.getVersion()));
+    }
+
+    public static @NonNull GetResult getResultFromProto(@NonNull GetResponse response) {
+        return new GetResult(
+                response.getValue().toByteArray(), getVersionFromProto(response.getVersion()));
+    }
+
+    public static @NonNull Version getVersionFromProto(@NonNull io.streamnative.oxia.proto.Version version) {
+        return new Version(
+                version.getVersionId(),
+                version.getCreatedTimestamp(),
+                version.getModifiedTimestamp(),
+                version.getModificationsCount(),
+                version.hasSessionId() ? Optional.of(version.getSessionId()) : Optional.empty(),
+                version.hasClientIdentity() ? Optional.of(version.getClientIdentity()) : Optional.empty());
     }
 }
